requirements.txt
aiogram==3.7.0
deep-translator==1.11.4
python-dotenv==1.0.0